+++ kotlintest-core/src/main/kotlin/io/kotlintest/specs/AbstractBehaviorSpec.kt
@@ -2,7 +2,9 @@ package io.kotlintest.specs
 
 import io.kotlintest.AbstractSpec
 import io.kotlintest.TestCase
-import io.kotlintest.TestScope
+import io.kotlintest.TestContainer
+import io.kotlintest.TestContext
+import io.kotlintest.lineNumber
 
 @Suppress("FunctionName")
 abstract class AbstractBehaviorSpec(body: AbstractBehaviorSpec.() -> Unit = {}) : AbstractSpec() {
@@ -13,31 +15,34 @@ abstract class AbstractBehaviorSpec(body: AbstractBehaviorSpec.() -> Unit = {}) 
 
   final override fun isInstancePerTest(): Boolean = false
 
-  fun Given(desc: String, init: GivenScope.() -> Unit) = given(desc, init)
-  fun given(desc: String, init: GivenScope.() -> Unit) {
-    rootScope.addContainer("Given $desc", this@AbstractBehaviorSpec, ::GivenScope, init)
+  fun Given(desc: String, init: GivenContext.() -> Unit) = given(desc, init)
+  fun given(desc: String, init: GivenContext.() -> Unit) {
+    rootScopes.add(TestContainer("Given $desc", this@AbstractBehaviorSpec, { GivenContext(it).init() }))
   }
 
-  inner class GivenScope : TestScope() {
+  inner class GivenContext(val context: TestContext) {
 
-    fun and(desc: String, init: GivenScope.() -> Unit) {
-      addContainer("And $desc", this@AbstractBehaviorSpec, ::GivenScope, init)
+    fun and(desc: String, init: GivenContext.() -> Unit) {
+      context.addScope(TestContainer("And $desc", this@AbstractBehaviorSpec, { GivenContext(it).init() }))
     }
 
-    fun When(desc: String, init: WhenScope.() -> Unit) = `when`(desc, init)
-    fun `when`(desc: String, init: WhenScope.() -> Unit) {
-      addContainer("When $desc", this@AbstractBehaviorSpec, ::WhenScope, init)
+    fun When(desc: String, init: WhenContext.() -> Unit) = `when`(desc, init)
+    fun `when`(desc: String, init: WhenContext.() -> Unit) {
+      context.addScope(TestContainer("When $desc", this@AbstractBehaviorSpec, { WhenContext(it).init() }))
     }
   }
 
-  inner class WhenScope : TestScope() {
+  inner class WhenContext(val context: TestContext) {
 
-    fun and(desc: String, init: WhenScope.() -> Unit) {
-      addContainer("And $desc", this@AbstractBehaviorSpec, ::WhenScope, init)
+    fun and(desc: String, init: WhenContext.() -> Unit) {
+      context.addScope(TestContainer("And $desc", this@AbstractBehaviorSpec, { WhenContext(it).init() }))
     }
 
-    fun Then(desc: String, test: () -> Unit): TestCase = then(desc, test)
-    fun then(desc: String, test: () -> Unit): TestCase =
-        addTest("Then $desc", this@AbstractBehaviorSpec, test, defaultTestCaseConfig)
+    fun Then(desc: String, test: TestContext.() -> Unit): TestCase = then(desc, test)
+    fun then(desc: String, test: TestContext.() -> Unit): TestCase {
+      val tc = TestCase("Then $desc", this@AbstractBehaviorSpec, test, lineNumber(), defaultTestCaseConfig)
+      context.addScope(tc)
+      return tc
+    }
   }
 }