+++ kotlintest-core/src/main/kotlin/io/kotlintest/specs/AnnotationSpec.kt
@@ -3,7 +3,7 @@ package io.kotlintest.specs
 import io.kotlintest.AbstractSpec
 import io.kotlintest.TestCase
 import io.kotlintest.TestContainer
-import io.kotlintest.TestScope
+import io.kotlintest.lineNumber
 
 annotation class Test
 
@@ -15,8 +15,8 @@ open class AbstractAnnotationSpec(body: AbstractAnnotationSpec.() -> Unit = {}) 
 
   override fun root(): TestContainer {
     val tests = javaClass.methods.filter { it.isAnnotationPresent(Test::class.java) }.map {
-      TestCase(it.name, this@AbstractAnnotationSpec, { it.invoke(this@AbstractAnnotationSpec) }, TestScope.lineNumber(), defaultTestCaseConfig)
+      TestCase(it.name, this@AbstractAnnotationSpec, { it.invoke(this@AbstractAnnotationSpec) }, lineNumber(), defaultTestCaseConfig)
     }
-    return TestContainer(name(), this, { tests }, true)
+    return TestContainer(name(), this, { context -> tests.forEach { context.addScope(it) } }, true)
   }
 }