+++ kotlintest-runner/kotlintest-runner-junit5/src/main/kotlin/io/kotlintest/runner/junit5/TestDiscovery.kt
@@ -41,9 +41,9 @@ object TestDiscovery {
       val spec: Spec = it.createInstance()
       val descriptor = TestContainerDescriptor.fromTestContainer(root.uniqueId, spec.root())
       root.addChild(descriptor)
-      root.sortChildren()
     }
 
+    root.sortChildren()
     return root
   }
 }